requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
black==24.1a1
certifi==2023.11.17
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.108.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-telegram-bot==20.7
sniffio==1.3.0
starlette==0.32.0.post1
tomli==2.0.1
typing_extensions==4.9.0
uvicorn==0.25.0
yarl==1.9.4
